+++ backend/cwa_lib/routers/doc_tasks.py
@@ -1,9 +1,8 @@
 from fastapi import APIRouter, Depends, HTTPException
-from common.enums.gvdbs_cfg_json import GVDBsCfgJSON
 from common.sql_db_async import AsyncSession, async_get_session
 from common.sql_models import User
 from cwa_lib.app import current_active_user
-from cwa_lib.pages.query_documents import QueryDocumentsOptions
+from cwa_lib.pages.query_documents import QueryDocumentsPage, QueryDocumentsOptions
 from cwa_lib.pydantic_schemas.doc_tasks import (
     DocTaskCreate, DocTaskQueryResult, DocTaskQueryShort, DocTaskDeleteResult, DocTaskOptionsResult
 )
@@ -19,33 +18,9 @@ async def create_task(
     user: User = Depends(current_active_user), 
     session: AsyncSession = Depends(async_get_session)
 ):
-    if not payload.doc_task_id:
-        result = await DocTasksTable(session).add_one(
-            group_id=user.group_id, 
-            user_id=user.user_id, 
-            gvdbs_id=payload.gvdbs_id,
-            gvdbs_cfg_json=GVDBsCfgJSON.from_dict(payload.gvdbs_cfg_json).as_dict(),
-            gllms_id=payload.gllms_id,
-            gc_id=payload.gc_id,
-            short_name=payload.short_name, 
-            input_text=payload.input_text, 
-            optional_text=payload.optional_text
-        )
-    else:
-        result = await DocTasksTable(session).add_second(
-            doc_task_id=payload.doc_task_id,
-            user_group_id=user.group_id, 
-            gvdbs_id=payload.gvdbs_id,
-            gvdbs_cfg_json=GVDBsCfgJSON.from_dict(payload.gvdbs_cfg_json).as_dict(),
-            gllms_id=payload.gllms_id,
-            gc_id=payload.gc_id,
-            short_name=payload.short_name, 
-            input_text=payload.input_text, 
-            optional_text=payload.optional_text
-        )
-    
-    if result is None:
-        raise HTTPException(status_code=404, detail="Cannot add new DocTask")
+    result = await QueryDocumentsPage(session, user).create_task(payload)    
+    if isinstance(result, str):
+        raise HTTPException(status_code=500, detail=result)
     return result
 
 
@@ -58,11 +33,11 @@ async def get_task(
     ):
     task = await DocTasksTable(session).query_one_by_doc_task_id(doc_task_id=doc_task_id)
     if task is None:
-        raise HTTPException(status_code=404, detail="DocTask not found")
+        raise HTTPException(status_code=500, detail="DocTask not found")
     # allow tasks only for user's group_id.
     # allow all tasks for superuser
     if (not user.is_superuser) and (task.group_id != user.group_id):
-        raise HTTPException(status_code=404, detail="DocTask not found")
+        raise HTTPException(status_code=404)
     if not user.is_superuser:  # hide context_json for non superusers
         task.context_json = None
     return task